pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "kuf-messdaten-excel-report"
authors = [
    {name = "Stefan Scheible", email = "stefan.scheible@kurz-fischer.com"},
]
version = "0.0.5"  # Required
description = "Dauerauswertung via python"

requires-python = ">=3.9"
keywords = ["one", "two"]
license = {text = "BSD-3-Clause"}
classifiers = [
]
dependencies = [
    "pandas",
    "psycopg2",
    "python-dotenv",
    "sqlalchemy",
    "xlsxwriter"
]


[tool.setuptools.packages.find]
# scanning for namespace packages is true by default in pyproject.toml, so
# you need NOT include the following line.
namespaces = true
where = ["src"]

[tool.setuptools]
# If there are data files included in your packages that need to be
# installed, specify them here.
package-data = {"kuf_messdaten_excel_report.resources" = ["*.jpg"]}



